Commit crimes setting.
This is the commit that I've been waiting for! Now there's a reason to
come back, and specifically to come back often. Stamina is already being
restored behind the scenes. Now I can spend that stamina on crimes to boost
my XP.

Crimes still need rewards aside from XP, and some custom messages so that
it's not so dry. Feels good though!